third_party/blink/web_tests/svg/custom/global-constructors-expected.txt

Test to make sure we expose all the global constructor objects requested by http://www.w3.org/TR/SVG11/ecmascript-binding.html

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS SVGElement.toString() is 'function SVGElement() { [native code] }'
PASS SVGAnimatedBoolean.toString() is 'function SVGAnimatedBoolean() { [native code] }'
PASS SVGAnimatedString.toString() is 'function SVGAnimatedString() { [native code] }'
PASS SVGStringList.toString() is 'function SVGStringList() { [native code] }'
PASS SVGAnimatedEnumeration.toString() is 'function SVGAnimatedEnumeration() { [native code] }'
PASS SVGAnimatedInteger.toString() is 'function SVGAnimatedInteger() { [native code] }'
PASS SVGNumber.toString() is 'function SVGNumber() { [native code] }'
PASS SVGAnimatedNumber.toString() is 'function SVGAnimatedNumber() { [native code] }'
PASS SVGNumberList.toString() is 'function SVGNumberList() { [native code] }'
PASS SVGAnimatedNumberList.toString() is 'function SVGAnimatedNumberList() { [native code] }'
PASS SVGLength.toString() is 'function SVGLength() { [native code] }'
PASS SVGAnimatedLength.toString() is 'function SVGAnimatedLength() { [native code] }'
PASS SVGLengthList.toString() is 'function SVGLengthList() { [native code] }'
PASS SVGAnimatedLengthList.toString() is 'function SVGAnimatedLengthList() { [native code] }'
PASS SVGAngle.toString() is 'function SVGAngle() { [native code] }'
PASS SVGAnimatedAngle.toString() is 'function SVGAnimatedAngle() { [native code] }'
PASS SVGRect.toString() is 'function SVGRect() { [native code] }'
PASS SVGAnimatedRect.toString() is 'function SVGAnimatedRect() { [native code] }'
FAIL SVGStylable.toString() should be function SVGStylable() { [native code] }. Threw exception ReferenceError: SVGStylable is not defined
FAIL SVGLocatable.toString() should be function SVGLocatable() { [native code] }. Threw exception ReferenceError: SVGLocatable is not defined
FAIL SVGTransformable.toString() should be function SVGTransformable() { [native code] }. Threw exception ReferenceError: SVGTransformable is not defined
FAIL SVGTests.toString() should be function SVGTests() { [native code] }. Threw exception ReferenceError: SVGTests is not defined
FAIL SVGURIReference.toString() should be function SVGURIReference() { [native code] }. Threw exception ReferenceError: SVGURIReference is not defined
FAIL SVGCSSRule.toString() should be function SVGCSSRule() { [native code] }. Threw exception ReferenceError: SVGCSSRule is not defined
PASS SVGSVGElement.toString() is 'function SVGSVGElement() { [native code] }'
PASS SVGGElement.toString() is 'function SVGGElement() { [native code] }'
PASS SVGDefsElement.toString() is 'function SVGDefsElement() { [native code] }'
PASS SVGDescElement.toString() is 'function SVGDescElement() { [native code] }'
FAIL SVGDiscardElement.toString() should be function SVGDiscardElement() { [native code] }. Threw exception ReferenceError: SVGDiscardElement is not defined
PASS SVGTitleElement.toString() is 'function SVGTitleElement() { [native code] }'
PASS SVGSymbolElement.toString() is 'function SVGSymbolElement() { [native code] }'
PASS SVGUseElement.toString() is 'function SVGUseElement() { [native code] }'
PASS SVGImageElement.toString() is 'function SVGImageElement() { [native code] }'
PASS SVGSwitchElement.toString() is 'function SVGSwitchElement() { [native code] }'
PASS SVGStyleElement.toString() is 'function SVGStyleElement() { [native code] }'
PASS SVGPoint.toString() is 'function SVGPoint() { [native code] }'
PASS SVGPointList.toString() is 'function SVGPointList() { [native code] }'
PASS SVGMatrix.toString() is 'function SVGMatrix() { [native code] }'
PASS SVGTransform.toString() is 'function SVGTransform() { [native code] }'
PASS SVGTransformList.toString() is 'function SVGTransformList() { [native code] }'
PASS SVGAnimatedTransformList.toString() is 'function SVGAnimatedTransformList() { [native code] }'
PASS SVGPreserveAspectRatio.toString() is 'function SVGPreserveAspectRatio() { [native code] }'
PASS SVGAnimatedPreserveAspectRatio.toString() is 'function SVGAnimatedPreserveAspectRatio() { [native code] }'
PASS SVGPathElement.toString() is 'function SVGPathElement() { [native code] }'
PASS SVGRectElement.toString() is 'function SVGRectElement() { [native code] }'
PASS SVGCircleElement.toString() is 'function SVGCircleElement() { [native code] }'
PASS SVGEllipseElement.toString() is 'function SVGEllipseElement() { [native code] }'
PASS SVGLineElement.toString() is 'function SVGLineElement() { [native code] }'
FAIL SVGAnimatedPoints.toString() should be function SVGAnimatedPoints() { [native code] }. Threw exception ReferenceError: SVGAnimatedPoints is not defined
PASS SVGPolylineElement.toString() is 'function SVGPolylineElement() { [native code] }'
PASS SVGPolygonElement.toString() is 'function SVGPolygonElement() { [native code] }'
PASS SVGTextContentElement.toString() is 'function SVGTextContentElement() { [native code] }'
PASS SVGTextPositioningElement.toString() is 'function SVGTextPositioningElement() { [native code] }'
PASS SVGTextElement.toString() is 'function SVGTextElement() { [native code] }'
PASS SVGTSpanElement.toString() is 'function SVGTSpanElement() { [native code] }'
PASS SVGTextPathElement.toString() is 'function SVGTextPathElement() { [native code] }'
PASS SVGMarkerElement.toString() is 'function SVGMarkerElement() { [native code] }'
FAIL SVGColorProfileElement.toString() should be function SVGColorProfileElement() { [native code] }. Threw exception ReferenceError: SVGColorProfileElement is not defined
FAIL SVGColorProfileRule.toString() should be function SVGColorProfileRule() { [native code] }. Threw exception ReferenceError: SVGColorProfileRule is not defined
PASS SVGGradientElement.toString() is 'function SVGGradientElement() { [native code] }'
PASS SVGLinearGradientElement.toString() is 'function SVGLinearGradientElement() { [native code] }'
PASS SVGRadialGradientElement.toString() is 'function SVGRadialGradientElement() { [native code] }'
PASS SVGStopElement.toString() is 'function SVGStopElement() { [native code] }'
PASS SVGPatternElement.toString() is 'function SVGPatternElement() { [native code] }'
PASS SVGClipPathElement.toString() is 'function SVGClipPathElement() { [native code] }'
PASS SVGMaskElement.toString() is 'function SVGMaskElement() { [native code] }'
PASS SVGFilterElement.toString() is 'function SVGFilterElement() { [native code] }'
FAIL SVGFilterPrimitiveStandardAttributes.toString() should be function SVGFilterPrimitiveStandardAttributes() { [native code] }. Threw exception ReferenceError: SVGFilterPrimitiveStandardAttributes is not defined
PASS SVGFEBlendElement.toString() is 'function SVGFEBlendElement() { [native code] }'
PASS SVGFEColorMatrixElement.toString() is 'function SVGFEColorMatrixElement() { [native code] }'
PASS SVGFEComponentTransferElement.toString() is 'function SVGFEComponentTransferElement() { [native code] }'
PASS SVGComponentTransferFunctionElement.toString() is 'function SVGComponentTransferFunctionElement() { [native code] }'
PASS SVGFEFuncRElement.toString() is 'function SVGFEFuncRElement() { [native code] }'
PASS SVGFEFuncGElement.toString() is 'function SVGFEFuncGElement() { [native code] }'
PASS SVGFEFuncBElement.toString() is 'function SVGFEFuncBElement() { [native code] }'
PASS SVGFEFuncAElement.toString() is 'function SVGFEFuncAElement() { [native code] }'
PASS SVGFECompositeElement.toString() is 'function SVGFECompositeElement() { [native code] }'
PASS SVGFEConvolveMatrixElement.toString() is 'function SVGFEConvolveMatrixElement() { [native code] }'
PASS SVGFEDiffuseLightingElement.toString() is 'function SVGFEDiffuseLightingElement() { [native code] }'
PASS SVGFEDistantLightElement.toString() is 'function SVGFEDistantLightElement() { [native code] }'
PASS SVGFEPointLightElement.toString() is 'function SVGFEPointLightElement() { [native code] }'
PASS SVGFESpotLightElement.toString() is 'function SVGFESpotLightElement() { [native code] }'
PASS SVGFEDisplacementMapElement.toString() is 'function SVGFEDisplacementMapElement() { [native code] }'
PASS SVGFEFloodElement.toString() is 'function SVGFEFloodElement() { [native code] }'
PASS SVGFEGaussianBlurElement.toString() is 'function SVGFEGaussianBlurElement() { [native code] }'
PASS SVGFEImageElement.toString() is 'function SVGFEImageElement() { [native code] }'
PASS SVGFEMergeElement.toString() is 'function SVGFEMergeElement() { [native code] }'
PASS SVGFEMergeNodeElement.toString() is 'function SVGFEMergeNodeElement() { [native code] }'
PASS SVGFEMorphologyElement.toString() is 'function SVGFEMorphologyElement() { [native code] }'
PASS SVGFEOffsetElement.toString() is 'function SVGFEOffsetElement() { [native code] }'
PASS SVGFESpecularLightingElement.toString() is 'function SVGFESpecularLightingElement() { [native code] }'
PASS SVGFETileElement.toString() is 'function SVGFETileElement() { [native code] }'
PASS SVGFETurbulenceElement.toString() is 'function SVGFETurbulenceElement() { [native code] }'
FAIL SVGCursorElement.toString() should be function SVGCursorElement() { [native code] }. Threw exception ReferenceError: SVGCursorElement is not defined
PASS SVGAElement.toString() is 'function SVGAElement() { [native code] }'
PASS SVGViewElement.toString() is 'function SVGViewElement() { [native code] }'
PASS SVGScriptElement.toString() is 'function SVGScriptElement() { [native code] }'
FAIL SVGEvent.toString() should be function SVGEvent() { [native code] }. Threw exception ReferenceError: SVGEvent is not defined
PASS SVGAnimationElement.toString() is 'function SVGAnimationElement() { [native code] }'
PASS SVGAnimateElement.toString() is 'function SVGAnimateElement() { [native code] }'
PASS SVGSetElement.toString() is 'function SVGSetElement() { [native code] }'
PASS SVGAnimateMotionElement.toString() is 'function SVGAnimateMotionElement() { [native code] }'
PASS SVGMPathElement.toString() is 'function SVGMPathElement() { [native code] }'
FAIL SVGAnimateColorElement.toString() should be function SVGAnimateColorElement() { [native code] }. Threw exception ReferenceError: SVGAnimateColorElement is not defined
PASS SVGAnimateTransformElement.toString() is 'function SVGAnimateTransformElement() { [native code] }'
FAIL SVGDefinitionSrcElement.toString() should be function SVGDefinitionSrcElement() { [native code] }. Threw exception ReferenceError: SVGDefinitionSrcElement is not defined
PASS SVGMetadataElement.toString() is 'function SVGMetadataElement() { [native code] }'
PASS SVGForeignObjectElement.toString() is 'function SVGForeignObjectElement() { [native code] }'
PASS successfullyParsed is true

TEST COMPLETE
